fix: move shutdown to settings
