+++ frontend/src/components/layouts/AppLayout.tsx
@@ -1,17 +1,15 @@
 import {
   Cable,
+  EllipsisVertical,
   ExternalLinkIcon,
   FlaskRound,
   LogOut,
   Menu, MessageCircleQuestion,
-  Power,
   Settings,
   Store,
-  Wallet
+  Wallet,
 } from "lucide-react";
 
-import { AlertDialogTrigger } from "@radix-ui/react-alert-dialog";
-import { CaretUpIcon } from "@radix-ui/react-icons";
 import React from "react";
 import {
   Link,
@@ -21,7 +19,6 @@ import {
   useNavigate,
 } from "react-router-dom";
 import SidebarHint from "src/components/SidebarHint";
-import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle } from "src/components/ui/alert-dialog";
 import { Avatar, AvatarFallback, AvatarImage } from "src/components/ui/avatar";
 import { Button } from "src/components/ui/button";
 import {
@@ -29,9 +26,8 @@ import {
   DropdownMenuContent,
   DropdownMenuGroup,
   DropdownMenuItem,
-  DropdownMenuLabel,
   DropdownMenuSeparator,
-  DropdownMenuTrigger,
+  DropdownMenuTrigger
 } from "src/components/ui/dropdown-menu";
 import { Sheet, SheetContent, SheetTrigger } from "src/components/ui/sheet";
 import { useToast } from "src/components/ui/use-toast";
@@ -74,48 +70,24 @@ export default function AppLayout() {
     toast({ title: "You are now logged out." });
   }, [csrf, navigate, toast]);
 
-  const shutdown = React.useCallback(async () => {
-    if (!csrf) {
-      throw new Error("csrf not loaded");
-    }
-
-    // TODO: Implement
-    // await request("/api/logout", {
-    //   method: "POST",
-    //   headers: {
-    //     "X-CSRF-Token": csrf,
-    //     "Content-Type": "application/json",
-    //   },
-    // });
-
-    // navigate("/", { replace: true });
-    // toast({ title: "You are now logged out." });
-
-  }, [csrf, navigate, toast]);
-
   function UserMenuContent() {
     return (
       <DropdownMenuContent align="end">
-        <DropdownMenuLabel>My Account</DropdownMenuLabel>
         <DropdownMenuGroup>
           <DropdownMenuItem>
             <ExternalLink
-              to="https://getalby.com/dashboard"
+              to="https://getalby.com/settings"
               className="w-full flex flex-row items-center gap-2"
             >
               <ExternalLinkIcon className="w-4 h-4" />
-              Go to getalby.com
+              Alby Account Settings
             </ExternalLink>
           </DropdownMenuItem>
         </DropdownMenuGroup>
         <DropdownMenuSeparator />
         <DropdownMenuItem onClick={logout}>
-          <LogOut className="w-4 h-4 mr-2" />
+          <LogOut className="w-4 h-4" />
           Log out
-        </DropdownMenuItem>
-        <DropdownMenuItem onClick={logout}>
-          <Power className="w-4 h-4 mr-2" />
-          Shutdown
         </DropdownMenuItem>
       </DropdownMenuContent>
     );
@@ -207,49 +179,14 @@ export default function AppLayout() {
                       {albyMe?.name || albyMe?.email}
                     </Link>
                   </div>
-
-                  <AlertDialog>
-                    <DropdownMenu>
-                      <DropdownMenuTrigger asChild>
-                        <Button variant="ghost" size="icon">
-                          <CaretUpIcon />
-                        </Button>
-                      </DropdownMenuTrigger>
-                      <UserMenuContent />
-                    </DropdownMenu>
-                    <DropdownMenu>
-                      <DropdownMenuTrigger asChild>
-                        <Button variant="ghost" size="icon">
-                          <Power className="w-4 h-4" />
-                        </Button>
-                      </DropdownMenuTrigger>
-                      <DropdownMenuContent>
-                        <DropdownMenuItem onClick={logout}>
-                          <LogOut className="w-4 h-4 mr-2" />
-                          Log out
-                        </DropdownMenuItem>
-                        <AlertDialogTrigger asChild>
-                          <DropdownMenuItem>
-                            <Power className="w-4 h-4 mr-2" />
-                            Shutdown
-                          </DropdownMenuItem>
-                        </AlertDialogTrigger>
-                      </DropdownMenuContent>
-                    </DropdownMenu>
-                    <AlertDialogContent>
-                      <AlertDialogHeader>
-                        <AlertDialogTitle>Do you want to turn off Alby Hub?</AlertDialogTitle>
-                        <AlertDialogDescription>
-                          This will turn off your Alby Hub and make your node offline. You won't be able to send or receive bitcoin until you unlock it.
-                        </AlertDialogDescription>
-                      </AlertDialogHeader>
-                      <AlertDialogFooter>
-                        <AlertDialogCancel>Cancel</AlertDialogCancel>
-                        <AlertDialogAction onClick={shutdown}>Continue</AlertDialogAction>
-                      </AlertDialogFooter>
-                    </AlertDialogContent>
-                  </AlertDialog>
-
+                  <DropdownMenu>
+                    <DropdownMenuTrigger asChild>
+                      <Button variant="ghost" size="icon">
+                        <EllipsisVertical className="w-4 h-4" />
+                      </Button>
+                    </DropdownMenuTrigger>
+                    <UserMenuContent />
+                  </DropdownMenu>
                 </div>
               </div>
             </div>
@@ -338,5 +275,3 @@ const MenuItem = ({
     {children}
   </NavLink>
 );
-
-MenuItem;
