+++ frontend/src/components/layouts/SettingsLayout.tsx
@@ -1,15 +1,63 @@
-import { ExternalLink } from "lucide-react";
-import { NavLink, Outlet } from "react-router-dom";
+import { ExternalLink, Power } from "lucide-react";
+import React from "react";
+import { NavLink, Outlet, useNavigate } from "react-router-dom";
 import AppHeader from "src/components/AppHeader";
-import { buttonVariants } from "src/components/ui/button";
+import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle, AlertDialogTrigger } from "src/components/ui/alert-dialog";
+import { Button, buttonVariants } from "src/components/ui/button";
+import { useToast } from "src/components/ui/use-toast";
+import { useCSRF } from "src/hooks/useCSRF";
 import { cn } from "src/lib/utils";
+import { request } from "src/utils/request";
 
 export default function SettingsLayout() {
+  const { data: csrf } = useCSRF();
+  const navigate = useNavigate();
+  const { toast } = useToast();
+
+  const shutdown = React.useCallback(async () => {
+    if (!csrf) {
+      throw new Error("csrf not loaded");
+    }
+
+    await request("/api/stop", {
+      method: "POST",
+      headers: {
+        "X-CSRF-Token": csrf,
+        "Content-Type": "application/json",
+      },
+    });
+
+    navigate("/", { replace: true });
+    toast({ title: "Your node has been turned off." });
+
+  }, [csrf, navigate, toast]);
+
   return (
     <>
       <AppHeader
         title="Settings"
-        description="Manage your account settings."
+        description="Manage your Alby Hub settings."
+        contentRight={
+          <AlertDialog>
+            <AlertDialogTrigger asChild>
+              <Button variant="destructive" size="icon">
+                <Power className="w-4 h-4" />
+              </Button>
+            </AlertDialogTrigger>
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                <AlertDialogTitle>Do you want to turn off Alby Hub?</AlertDialogTitle>
+                <AlertDialogDescription>
+                  This will turn off your Alby Hub and make your node offline. You won't be able to send or receive bitcoin until you unlock it.
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+              <AlertDialogFooter>
+                <AlertDialogCancel>Cancel</AlertDialogCancel>
+                <AlertDialogAction onClick={shutdown}>Continue</AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
+        }
       />
       <div className="flex flex-col space-y-8 lg:flex-row lg:space-x-12 lg:space-y-0">
         <aside className="-mx-4 lg:w-1/5">
